+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/constants/JsonFormConstants.java
@@ -32,7 +32,7 @@ public class JsonFormConstants {
     public static final String NAME = "name";
     public static final String TYPE = "type";
     public static final String READ_ONLY = "read_only";
-    public static final String DISABLED = "disabled";
+    public static final String EDITABLE = "editable";
     public static final String BEHAVIOUR_FINISH_FORM = "finish_form";
     public static final String BEHAVIOUR_NEXT_STEP = "next_step";
     public static final String HINT = "hint";
@@ -58,9 +58,10 @@ public class JsonFormConstants {
     public static final String V_NUMERIC_INTEGER = "v_numeric_integer";
     public static final String V_MIN = "v_min";
     public static final String V_MAX = "v_max";
-    public static final String NUMBER_SELECTORS = "numbers_selector";
+    public static final String NUMBERS_SELECTOR = "numbers_selector";
     public static final String START_SELECTION_NUMBER = "start_number";
     public static final String NUMBER_OF_SELECTORS = "number_of_selectors";
+    public static final String NUMBER_OF_SELECTORS_ORIGINAL = "number_of_selectors_original";
     public static final String MAX_SELECTION_VALUE = "max_value";
     public static final String EDIT_TYPE = "edit_type";
     public static final String TEXT = "text";
@@ -122,6 +123,8 @@ public class JsonFormConstants {
     public static final String LEFT_MARGIN = "left_margin";
     public static final String UPLOAD_BUTTON_TEXT = "uploadButtonText";
     public static final String EXPANSION_PANEL = "expansion_panel";
+    public static final String JSON_OBJECT_KEY = "json_object_key";
+    public static final String IS_POPUP = "is_popup";
 
     public static class JSON_FORM_KEY {
         public static final String EX_RULES = "ex-rules";
@@ -145,5 +148,24 @@ public class JsonFormConstants {
 
     public static class EDIT_TEXT_TYPE {
         public static final String NUMBER = "number";
+    }
+
+    public static class INTENT_ACTION {
+        public static final String NUMBER_SELECTOR_FACTORY = "number_selector_factory";
+        public static final String JSON_FORM_ACTIVITY = "json_form_activity";
+    }
+
+    public static class SUFFIX {
+        public static final String SPINNER = "_spinner";
+        public static final String TEXT_VIEW = "_textview";
+    }
+
+    public static class INTENT_KEY {
+        public static final String MESSAGE = "message";
+        public static final String MESSAGE_TYPE = "message_type";
+    }
+
+    public static class MESSAGE_TYPE {
+        public static final String GLOBAL_VALUES = "global_values";
     }
 }