+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/utils/FormUtils.java
@@ -6,7 +6,6 @@ import android.app.FragmentTransaction;
 import android.content.Context;
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.support.v7.widget.AppCompatEditText;
 import android.support.v7.widget.AppCompatTextView;
 import android.text.Html;
 import android.text.TextUtils;
@@ -58,7 +57,8 @@ public class FormUtils {
     public static final int WRAP_CONTENT = -2;
     public static final String METADATA_PROPERTY = "metadata";
     public static final String LOOK_UP_JAVAROSA_PROPERTY = "look_up";
-    private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("dd-MM-yyyy");
+    public static final String NATIIVE_FORM_DATE_FORMAT_PATTERN = "dd-MM-yyyy";
+    private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat(NATIIVE_FORM_DATE_FORMAT_PATTERN);
     private static final SimpleDateFormat DATE_TIME_FORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
     private static final String START_JAVAROSA_PROPERTY = "start";
     private static final String END_JAVAROSA_PROPERTY = "end";
@@ -455,17 +455,22 @@ public class FormUtils {
         }
     }
 
-    public static void setEditMode(JSONObject jsonObject, AppCompatEditText editText, ImageView editButton) throws JSONException {
-        if (jsonObject.has(JsonFormConstants.DISABLED) || (jsonObject.has(JsonFormConstants.DISABLED)
-                && jsonObject.has(JsonFormConstants.READ_ONLY))) {
-            boolean disabled = jsonObject.getBoolean(JsonFormConstants.DISABLED);
-            editText.setEnabled(!disabled);
-            editText.setFocusable(!disabled);
-            editButton.setVisibility(View.GONE);
+    public static void setEditMode(JSONObject jsonObject, View editableView, ImageView editButton) throws JSONException {
+        if (jsonObject.has(JsonFormConstants.EDITABLE)) {
+            boolean editable = jsonObject.getBoolean(JsonFormConstants.EDITABLE);
+            if (editable) {
+                editButton.setVisibility(View.VISIBLE);
+                editableView.setEnabled(false);
+            } else {
+                editButton.setVisibility(View.GONE);
+            }
         } else if (jsonObject.has(JsonFormConstants.READ_ONLY)) {
             boolean readyOnly = jsonObject.getBoolean(JsonFormConstants.READ_ONLY);
-            editText.setEnabled(!readyOnly);
+            editableView.setEnabled(!readyOnly);
+            editButton.setVisibility(View.GONE);
+        } else if (jsonObject.has(JsonFormConstants.EDITABLE) && jsonObject.has(JsonFormConstants.READ_ONLY)) {
             editButton.setVisibility(View.VISIBLE);
+            editableView.setEnabled(false);
         }
     }