+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/utils/Utils.java
@@ -1,10 +1,134 @@
 package com.vijay.jsonwizard.utils;
 
 import android.content.Context;
+import android.support.v4.util.TimeUtils;
+import android.text.TextUtils;
+import android.util.Log;
 import android.widget.Toast;
 
+import com.vijay.jsonwizard.widgets.DatePickerFactory;
+
+import java.text.ParseException;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.concurrent.TimeUnit;
+
 public class Utils {
+
+    private static final String TAG = Utils.class.getCanonicalName();
+
     public static void showToast(Context context, String message) {
         Toast.makeText(context, message, Toast.LENGTH_LONG).show();
+    }
+
+    public static Date getDateFromString(String dtStart) {
+        try {
+            Date date = DatePickerFactory.DATE_FORMAT.parse(dtStart);
+            return date;
+        } catch (ParseException e) {
+            Log.e(TAG, e.getMessage(), e);
+            return null;
+        }
+    }
+
+    public static String getStringFromDate(Date date) {
+        try {
+            return DatePickerFactory.DATE_FORMAT.format(date);
+        } catch (Exception e) {
+            Log.e(TAG, e.getMessage(), e);
+            return null;
+        }
+    }
+
+    public static String reverseDateString(String str, String delimiter) {
+        String[] strr = str.split(delimiter);
+        return strr[2] + "-" + strr[1] + "-" + strr[0];
+    }
+
+    public static String getDuration(String date) {
+        if (!TextUtils.isEmpty(date)) {
+            Calendar calendar = FormUtils.getDate(date);
+            calendar.set(Calendar.HOUR_OF_DAY, 0);
+            calendar.set(Calendar.MINUTE, 0);
+            calendar.set(Calendar.SECOND, 0);
+            calendar.set(Calendar.MILLISECOND, 0);
+
+            Calendar now = Calendar.getInstance();
+            now.set(Calendar.HOUR_OF_DAY, 0);
+            now.set(Calendar.MINUTE, 0);
+            now.set(Calendar.SECOND, 0);
+            now.set(Calendar.MILLISECOND, 0);
+
+            long timeDiff = Math.abs(now.getTimeInMillis() - calendar.getTimeInMillis());
+            StringBuilder builder = new StringBuilder();
+            TimeUtils.formatDuration(timeDiff, builder);
+            String duration = "";
+            if (timeDiff >= 0
+                    && timeDiff <= TimeUnit.MILLISECONDS.convert(13, TimeUnit.DAYS)) {
+                // Represent in days
+                long days = TimeUnit.DAYS.convert(timeDiff, TimeUnit.MILLISECONDS);
+                duration = days + "d";
+            } else if (timeDiff > TimeUnit.MILLISECONDS.convert(13, TimeUnit.DAYS)
+                    && timeDiff <= TimeUnit.MILLISECONDS.convert(97, TimeUnit.DAYS)) {
+                // Represent in weeks and days
+                int weeks = (int) Math.floor((float) timeDiff /
+                        TimeUnit.MILLISECONDS.convert(7, TimeUnit.DAYS));
+                int days = (int) Math.floor((float) (timeDiff -
+                        TimeUnit.MILLISECONDS.convert(weeks * 7, TimeUnit.DAYS)) /
+                        TimeUnit.MILLISECONDS.convert(1, TimeUnit.DAYS));
+
+                if (days >= 7) {
+                    days = 0;
+                    weeks++;
+                }
+
+                duration = weeks + "w";
+                if (days > 0) {
+                    duration += " " + days + "d";
+                }
+            } else if (timeDiff > TimeUnit.MILLISECONDS.convert(97, TimeUnit.DAYS)
+                    && timeDiff <= TimeUnit.MILLISECONDS.convert(363, TimeUnit.DAYS)) {
+                // Represent in months and weeks
+                int months = (int) Math.floor((float) timeDiff
+                        / TimeUnit.MILLISECONDS.convert(30, TimeUnit.DAYS));
+                int weeks = (int) Math.floor((float) (timeDiff - TimeUnit.MILLISECONDS.convert(
+                        months * 30, TimeUnit.DAYS)) /
+                        TimeUnit.MILLISECONDS.convert(7, TimeUnit.DAYS));
+
+                if (weeks >= 4) {
+                    weeks = 0;
+                    months++;
+                }
+
+                if (months < 12) {
+                    duration = months + "m";
+                    if (weeks > 0 && months < 12) {
+                        duration += " " + weeks + "w";
+                    }
+                } else if (months >= 12) {
+                    duration = "1y";
+                }
+            } else {
+                // Represent in years and months
+                int years = (int) Math.floor((float) timeDiff
+                        / TimeUnit.MILLISECONDS.convert(365, TimeUnit.DAYS));
+                int months = (int) Math.floor((float) (timeDiff -
+                        TimeUnit.MILLISECONDS.convert(years * 365, TimeUnit.DAYS)) /
+                        TimeUnit.MILLISECONDS.convert(30, TimeUnit.DAYS));
+
+                if (months >= 12) {
+                    months = 0;
+                    years++;
+                }
+
+                duration = years + "y";
+                if (months > 0) {
+                    duration += " " + months + "m";
+                }
+            }
+
+            return duration;
+        }
+        return null;
     }
 }